requirements.txt
google_api_python_client==2.19.1
python-dotenv==0.19.0
spotipy==2.19.0
